#87 Refactor: swagger 관련 파일 디렉토리 변경
