+++ src/main/java/com/gyeongditor/storyfield/Controller/ImageController.java
@@ -2,7 +2,7 @@ package com.gyeongditor.storyfield.Controller;
 
 import com.gyeongditor.storyfield.dto.ApiResponseDTO;
 import com.gyeongditor.storyfield.service.S3Service;
-import com.gyeongditor.storyfield.swagger.ImageApi;
+import com.gyeongditor.storyfield.swagger.api.ImageApi;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RestController;
