+++ src/main/java/com/gyeongditor/storyfield/Controller/UserController.java
@@ -5,9 +5,8 @@ import com.gyeongditor.storyfield.dto.UserDTO.SignUpDTO;
 import com.gyeongditor.storyfield.dto.UserDTO.UpdateUserDTO;
 import com.gyeongditor.storyfield.dto.UserDTO.UserResponseDTO;
 import com.gyeongditor.storyfield.service.UserService;
-import com.gyeongditor.storyfield.swagger.UserApi;
+import com.gyeongditor.storyfield.swagger.api.UserApi;
 import jakarta.servlet.http.HttpServletRequest;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RestController;
