+++ src/main/java/com/gyeongditor/storyfield/swagger/api/AuthApi.java
@@ -1,9 +1,11 @@
-package com.gyeongditor.storyfield.swagger;
+package com.gyeongditor.storyfield.swagger.api;
 
 import com.gyeongditor.storyfield.dto.ApiResponseDTO;
 import com.gyeongditor.storyfield.dto.UserDTO.LoginDTO;
 import com.gyeongditor.storyfield.response.ErrorCode;
 import com.gyeongditor.storyfield.response.SuccessCode;
+import com.gyeongditor.storyfield.swagger.config.ApiErrorResponse;
+import com.gyeongditor.storyfield.swagger.config.ApiSuccessResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
