+++ src/main/java/com/gyeongditor/storyfield/swagger/config/ApiSuccessExampleProcessor.java
@@ -1,4 +1,4 @@
-package com.gyeongditor.storyfield.swagger;
+package com.gyeongditor.storyfield.swagger.config;
 
 import com.gyeongditor.storyfield.response.SuccessCode;
 import io.swagger.v3.oas.models.examples.Example;
